+++ Simple Game Library/src/at/fhooe/mtd/sgl/SglInfo.java
@@ -14,7 +14,7 @@ public class SglInfo {
 
     public static final String NAME = "Simple Game Library";
     
-    public static final String VERSION = "1.4.1";
+    public static final String VERSION = "1.4.2 (preliminary)";
     
     public static final String COPYRIGHT_HOLDER 
         = "Roman Divotkey, Univ. of Applied Sciences Upper Austria.";